# improve tracestate performance