Tracestate is essential logic that cannot be skipped in some scenarios that use tracestate. In the past, the implementation of tracestate did not prioritize performance, using complex regular expressions, causing its efficiency to be low. In some scenarios, too much CPU was wasted.

For example, even if it doesn't hit sampling, startServerSpan consumes 25% of the entire request CPU.
![image](https://github.com/open-telemetry/opentelemetry-go/assets/571878/cf982a99-990f-4f28-b733-919f3e55f308)

Typical server span tracestate:
![image](https://github.com/open-telemetry/opentelemetry-go/assets/571878/357a94bd-09c1-45c1-8212-4a275c9c5068)

Typical client span Propagation
![image](https://github.com/open-telemetry/opentelemetry-go/assets/571878/d7b1bdad-9f66-452b-9e1f-aae09d286c17)

It can be seen that the poor performance of tracestate is mainly due to regular expressions and member.String.

This CR can significantly improve Propagation by optimizing it in the following ways:

Use string.Builder to directly construct the result
Reduce the redundant copying during Insert
Avoid using regex
All unit tests pass. Benchmarks: 

```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/trace
cpu: Intel(R) Xeon(R) Platinum 8255C CPU @ 2.50GHz
                                                                   │  old-ts.txt   │           65bd30b-ts.txt            │
                                                                   │    sec/op     │   sec/op     vs base                │
TraceStateParse/duplicate_with_the_same_value-16                      2072.0n ± 1%   469.8n ± 0%  -77.33% (p=0.000 n=10)
TraceStateParse/duplicate_with_different_values-16                    2084.0n ± 1%   470.2n ± 1%  -77.44% (p=0.000 n=10)
TraceStateParse/improperly_formatted_key/value_pair-16                1109.0n ± 1%   629.2n ± 0%  -43.26% (p=0.000 n=10)
TraceStateParse/upper_case_key-16                                      832.4n ± 2%   621.9n ± 0%  -25.29% (p=0.000 n=10)
TraceStateParse/key_with_invalid_character-16                         1100.0n ± 0%   629.8n ± 0%  -42.74% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_empty_tenant_key-16           1149.0n ± 1%   631.7n ± 1%  -45.02% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_only_tenant-16                 861.9n ± 1%   621.9n ± 0%  -27.85% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_double_tenant_separator-16    1148.0n ± 1%   634.1n ± 1%  -44.76% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_multiple_tenants-16           1268.0n ± 1%   642.5n ± 0%  -49.33% (p=0.000 n=10)
TraceStateParse/key_too_long-16                                       7196.0n ± 1%   906.4n ± 0%  -87.40% (p=0.000 n=10)
TraceStateParse/key_too_long,_with_tenant-16                         28267.0n ± 1%   902.7n ± 0%  -96.81% (p=0.000 n=10)
TraceStateParse/tenant_too_long-16                                    3297.5n ± 3%   784.5n ± 0%  -76.21% (p=0.000 n=10)
TraceStateParse/multiple_values_for_a_single_key-16                   1352.0n ± 1%   634.5n ± 0%  -53.07% (p=0.000 n=10)
TraceStateParse/no_value-16                                           1144.5n ± 1%   618.4n ± 0%  -45.96% (p=0.000 n=10)
TraceStateParse/too_many_members-16                                   37.155µ ± 2%   7.718µ ± 0%  -79.23% (p=0.000 n=10)
TraceStateParse/valid_key/value_list-16                               5547.5n ± 1%   265.9n ± 0%  -95.21% (p=0.000 n=10)
TraceStateParse/valid_key/value_list_with_tenant-16                   8599.0n ± 1%   285.8n ± 0%  -96.68% (p=0.000 n=10)
TraceStateParse/empty_input-16                                         2.289n ± 1%   2.322n ± 1%   +1.42% (p=0.001 n=10)
TraceStateParse/single_key_and_value-16                                944.5n ± 0%   139.3n ± 0%  -85.25% (p=0.000 n=10)
TraceStateParse/single_key_and_value_with_empty_separator-16           973.8n ± 0%   138.2n ± 0%  -85.80% (p=0.000 n=10)
TraceStateParse/multiple_keys_and_values-16                           1888.5n ± 1%   297.2n ± 0%  -84.26% (p=0.000 n=10)
TraceStateParse/with_a_key_at_maximum_length-16                      12882.5n ± 1%   588.9n ± 0%  -95.43% (p=0.000 n=10)
TraceStateParse/with_a_key_and_tenant_at_maximum_length-16           27831.5n ± 1%   598.8n ± 1%  -97.85% (p=0.000 n=10)
TraceStateParse/with_maximum_members-16                               34.634µ ± 1%   6.717µ ± 0%  -80.61% (p=0.000 n=10)
TraceStateParse/with_several_members-16                               4938.5n ± 1%   766.4n ± 0%  -84.48% (p=0.000 n=10)
TraceStateParse/with_tabs_between_members-16                          2954.5n ± 0%   491.6n ± 0%  -83.36% (p=0.000 n=10)
TraceStateParse/with_multiple_tabs_between_members-16                 2950.5n ± 2%   491.3n ± 0%  -83.35% (p=0.000 n=10)
TraceStateParse/with_space_at_the_end_of_the_member-16                 963.4n ± 1%   138.8n ± 0%  -85.60% (p=0.000 n=10)
TraceStateParse/with_tab_at_the_end_of_the_member-16                   935.5n ± 1%   138.1n ± 1%  -85.24% (p=0.000 n=10)
TraceStateParse/with_tab_and_space_at_the_end_of_the_member-16         974.9n ± 1%   139.1n ± 0%  -85.73% (p=0.000 n=10)
TraceStateString/valid_key/value_list-16                              280.10n ± 1%   69.69n ± 0%  -75.12% (p=0.000 n=10)
TraceStateString/valid_key/value_list_with_tenant-16                  283.35n ± 1%   72.46n ± 0%  -74.43% (p=0.000 n=10)
TraceStateString/single_key_and_value-16                              246.45n ± 1%   29.57n ± 0%  -88.00% (p=0.000 n=10)
TraceStateString/single_key_and_value_with_empty_separator-16         244.60n ± 1%   29.61n ± 0%  -87.90% (p=0.000 n=10)
TraceStateString/multiple_keys_and_values-16                          536.55n ± 1%   48.22n ± 0%  -91.01% (p=0.000 n=10)
TraceStateString/with_a_key_at_maximum_length-16                       667.9n ± 1%   109.9n ± 1%  -83.54% (p=0.000 n=10)
TraceStateString/with_a_key_and_tenant_at_maximum_length-16            666.9n ± 1%   109.9n ± 1%  -83.52% (p=0.000 n=10)
TraceStateString/with_maximum_members-16                              7215.0n ± 1%   439.2n ± 0%  -93.91% (p=0.000 n=10)
TraceStateString/with_several_members-16                             1196.00n ± 1%   88.43n ± 0%  -92.61% (p=0.000 n=10)
TraceStateString/with_tabs_between_members-16                         745.65n ± 1%   63.42n ± 2%  -91.49% (p=0.000 n=10)
TraceStateString/with_multiple_tabs_between_members-16                747.85n ± 1%   63.45n ± 0%  -91.52% (p=0.000 n=10)
TraceStateString/with_space_at_the_end_of_the_member-16               245.25n ± 1%   29.66n ± 0%  -87.91% (p=0.000 n=10)
TraceStateString/with_tab_at_the_end_of_the_member-16                 244.75n ± 2%   29.56n ± 0%  -87.92% (p=0.000 n=10)
TraceStateString/with_tab_and_space_at_the_end_of_the_member-16       242.75n ± 1%   29.67n ± 1%  -87.78% (p=0.000 n=10)
TraceStateInsert/add_new-16                                            916.2n ± 1%   122.3n ± 1%  -86.66% (p=0.000 n=10)
TraceStateInsert/replace-16                                            491.4n ± 1%   123.6n ± 0%  -74.85% (p=0.000 n=10)
TraceStateInsert/invalid_key-16                                       584.40n ± 1%   20.07n ± 1%  -96.56% (p=0.000 n=10)
TraceStateInsert/invalid_value-16                                     627.70n ± 0%   20.14n ± 0%  -96.79% (p=0.000 n=10)
TraceStateInsert/invalid_key/value-16                                 587.20n ± 2%   20.18n ± 1%  -96.56% (p=0.000 n=10)
TraceStateInsert/drop_the_right-most_member(oldest)_in_queue-16        778.4n ± 1%   383.2n ± 1%  -50.77% (p=0.000 n=10)
geomean                                                                1.266µ        199.5n       -84.24%

                                                                   │   old-ts.txt    │              65bd30b-ts.txt               │
                                                                   │      B/op       │     B/op      vs base                     │
TraceStateParse/duplicate_with_the_same_value-16                        387.0 ± 0%       128.0 ± 0%   -66.93% (p=0.000 n=10)
TraceStateParse/duplicate_with_different_values-16                      387.0 ± 0%       128.0 ± 0%   -66.93% (p=0.000 n=10)
TraceStateParse/improperly_formatted_key/value_pair-16                  226.0 ± 0%       208.0 ± 0%    -7.96% (p=0.000 n=10)
TraceStateParse/upper_case_key-16                                       226.0 ± 0%       208.0 ± 0%    -7.96% (p=0.000 n=10)
TraceStateParse/key_with_invalid_character-16                           226.0 ± 0%       208.0 ± 0%    -7.96% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_empty_tenant_key-16             242.0 ± 0%       208.0 ± 0%   -14.05% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_only_tenant-16                  243.0 ± 0%       208.0 ± 0%   -14.40% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_double_tenant_separator-16      242.0 ± 0%       208.0 ± 0%   -14.05% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_multiple_tenants-16             242.0 ± 0%       208.0 ± 0%   -14.05% (p=0.000 n=10)
TraceStateParse/key_too_long-16                                        1358.0 ± 0%       752.0 ± 0%   -44.62% (p=0.000 n=10)
TraceStateParse/key_too_long,_with_tenant-16                           1325.0 ± 0%       720.0 ± 0%   -45.66% (p=0.000 n=10)
TraceStateParse/tenant_too_long-16                                      614.0 ± 0%       256.0 ± 0%   -58.31% (p=0.000 n=10)
TraceStateParse/multiple_values_for_a_single_key-16                     226.0 ± 0%       208.0 ± 0%    -7.96% (p=0.000 n=10)
TraceStateParse/no_value-16                                             226.0 ± 0%       192.0 ± 0%   -15.04% (p=0.000 n=10)
TraceStateParse/too_many_members-16                                  10.471Ki ± 0%     6.193Ki ± 0%   -40.85% (p=0.000 n=10)
TraceStateParse/valid_key/value_list-16                                161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/valid_key/value_list_with_tenant-16                    161.00 ± 1%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/empty_input-16                                          0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
TraceStateParse/single_key_and_value-16                                161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/single_key_and_value_with_empty_separator-16           177.00 ± 0%       32.00 ± 0%   -81.92% (p=0.000 n=10)
TraceStateParse/multiple_keys_and_values-16                            355.00 ± 0%       96.00 ± 0%   -72.96% (p=0.000 n=10)
TraceStateParse/with_a_key_at_maximum_length-16                        355.00 ± 0%       96.00 ± 0%   -72.96% (p=0.000 n=10)
TraceStateParse/with_a_key_and_tenant_at_maximum_length-16             355.00 ± 0%       96.00 ± 0%   -72.96% (p=0.000 n=10)
TraceStateParse/with_maximum_members-16                               8.170Ki ± 0%     4.092Ki ± 0%   -49.92% (p=0.000 n=10)
TraceStateParse/with_several_members-16                                1129.5 ± 0%       480.0 ± 0%   -57.50% (p=0.000 n=10)
TraceStateParse/with_tabs_between_members-16                            613.0 ± 0%       224.0 ± 0%   -63.46% (p=0.000 n=10)
TraceStateParse/with_multiple_tabs_between_members-16                   613.0 ± 0%       224.0 ± 0%   -63.46% (p=0.000 n=10)
TraceStateParse/with_space_at_the_end_of_the_member-16                 161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/with_tab_at_the_end_of_the_member-16                   161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/with_tab_and_space_at_the_end_of_the_member-16         161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateString/valid_key/value_list-16                                192.0 ± 0%       144.0 ± 0%   -25.00% (p=0.000 n=10)
TraceStateString/valid_key/value_list_with_tenant-16                    208.0 ± 0%       160.0 ± 0%   -23.08% (p=0.000 n=10)
TraceStateString/single_key_and_value-16                               53.000 ± 0%       5.000 ± 0%   -90.57% (p=0.000 n=10)
TraceStateString/single_key_and_value_with_empty_separator-16          53.000 ± 0%       5.000 ± 0%   -90.57% (p=0.000 n=10)
TraceStateString/multiple_keys_and_values-16                           120.00 ± 0%       16.00 ± 0%   -86.67% (p=0.000 n=10)
TraceStateString/with_a_key_at_maximum_length-16                        677.0 ± 0%       288.0 ± 0%   -57.46% (p=0.000 n=10)
TraceStateString/with_a_key_and_tenant_at_maximum_length-16             677.0 ± 0%       288.0 ± 0%   -57.46% (p=0.000 n=10)
TraceStateString/with_maximum_members-16                               2081.0 ± 0%       288.0 ± 0%   -86.16% (p=0.000 n=10)
TraceStateString/with_several_members-16                               304.00 ± 0%       32.00 ± 0%   -89.47% (p=0.000 n=10)
TraceStateString/with_tabs_between_members-16                          184.00 ± 0%       24.00 ± 0%   -86.96% (p=0.000 n=10)
TraceStateString/with_multiple_tabs_between_members-16                 184.00 ± 0%       24.00 ± 0%   -86.96% (p=0.000 n=10)
TraceStateString/with_space_at_the_end_of_the_member-16                53.000 ± 0%       5.000 ± 0%   -90.57% (p=0.000 n=10)
TraceStateString/with_tab_at_the_end_of_the_member-16                  53.000 ± 0%       5.000 ± 0%   -90.57% (p=0.000 n=10)
TraceStateString/with_tab_and_space_at_the_end_of_the_member-16        53.000 ± 0%       5.000 ± 0%   -90.57% (p=0.000 n=10)
TraceStateInsert/add_new-16                                             291.0 ± 0%       128.0 ± 0%   -56.01% (p=0.000 n=10)
TraceStateInsert/replace-16                                             96.00 ± 1%       96.00 ± 0%         ~ (p=0.211 n=10)
TraceStateInsert/invalid_key-16                                         80.00 ± 0%        0.00 ± 0%  -100.00% (p=0.000 n=10)
TraceStateInsert/invalid_value-16                                       80.00 ± 0%        0.00 ± 0%  -100.00% (p=0.000 n=10)
TraceStateInsert/invalid_key/value-16                                   80.00 ± 0%        0.00 ± 0%  -100.00% (p=0.000 n=10)
TraceStateInsert/drop_the_right-most_member(oldest)_in_queue-16       1.012Ki ± 0%     1.000Ki ± 0%    -1.16% (p=0.000 n=10)
geomean                                                                            ²                 ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                                                                   │  old-ts.txt   │             65bd30b-ts.txt              │
                                                                   │   allocs/op   │ allocs/op   vs base                     │
TraceStateParse/duplicate_with_the_same_value-16                      8.000 ± 0%     3.000 ± 0%   -62.50% (p=0.000 n=10)
TraceStateParse/duplicate_with_different_values-16                    8.000 ± 0%     3.000 ± 0%   -62.50% (p=0.000 n=10)
TraceStateParse/improperly_formatted_key/value_pair-16                6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/upper_case_key-16                                     6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/key_with_invalid_character-16                         6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_empty_tenant_key-16           6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_only_tenant-16                6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_double_tenant_separator-16    6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_multiple_tenants-16           6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/key_too_long-16                                      14.000 ± 0%     6.000 ± 0%   -57.14% (p=0.000 n=10)
TraceStateParse/key_too_long,_with_tenant-16                         14.000 ± 0%     6.000 ± 0%   -57.14% (p=0.000 n=10)
TraceStateParse/tenant_too_long-16                                   14.000 ± 0%     6.000 ± 0%   -57.14% (p=0.000 n=10)
TraceStateParse/multiple_values_for_a_single_key-16                   6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/no_value-16                                           6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/too_many_members-16                                   80.00 ± 0%     13.00 ± 0%   -83.75% (p=0.000 n=10)
TraceStateParse/valid_key/value_list-16                               4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/valid_key/value_list_with_tenant-16                   4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/empty_input-16                                        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
TraceStateParse/single_key_and_value-16                               4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/single_key_and_value_with_empty_separator-16          4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/multiple_keys_and_values-16                           7.000 ± 0%     2.000 ± 0%   -71.43% (p=0.000 n=10)
TraceStateParse/with_a_key_at_maximum_length-16                       7.000 ± 0%     2.000 ± 0%   -71.43% (p=0.000 n=10)
TraceStateParse/with_a_key_and_tenant_at_maximum_length-16            7.000 ± 0%     2.000 ± 0%   -71.43% (p=0.000 n=10)
TraceStateParse/with_maximum_members-16                               75.00 ± 0%     10.00 ± 0%   -86.67% (p=0.000 n=10)
TraceStateParse/with_several_members-16                              15.000 ± 0%     4.000 ± 0%   -73.33% (p=0.000 n=10)
TraceStateParse/with_tabs_between_members-16                         10.000 ± 0%     3.000 ± 0%   -70.00% (p=0.000 n=10)
TraceStateParse/with_multiple_tabs_between_members-16                10.000 ± 0%     3.000 ± 0%   -70.00% (p=0.000 n=10)
TraceStateParse/with_space_at_the_end_of_the_member-16                4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/with_tab_at_the_end_of_the_member-16                  4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/with_tab_and_space_at_the_end_of_the_member-16        4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/valid_key/value_list-16                              4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/valid_key/value_list_with_tenant-16                  4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/single_key_and_value-16                              4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/single_key_and_value_with_empty_separator-16         4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/multiple_keys_and_values-16                          8.000 ± 0%     1.000 ± 0%   -87.50% (p=0.000 n=10)
TraceStateString/with_a_key_at_maximum_length-16                      8.000 ± 0%     1.000 ± 0%   -87.50% (p=0.000 n=10)
TraceStateString/with_a_key_and_tenant_at_maximum_length-16           8.000 ± 0%     1.000 ± 0%   -87.50% (p=0.000 n=10)
TraceStateString/with_maximum_members-16                             98.000 ± 0%     1.000 ± 0%   -98.98% (p=0.000 n=10)
TraceStateString/with_several_members-16                             17.000 ± 0%     1.000 ± 0%   -94.12% (p=0.000 n=10)
TraceStateString/with_tabs_between_members-16                        11.000 ± 0%     1.000 ± 0%   -90.91% (p=0.000 n=10)
TraceStateString/with_multiple_tabs_between_members-16               11.000 ± 0%     1.000 ± 0%   -90.91% (p=0.000 n=10)
TraceStateString/with_space_at_the_end_of_the_member-16               4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
…