@MrAlias 2023-11-16
The increases in allocations is concerning. The impact on GC is not captured in the benchmarks but will be impactful to real systems.