@xiehuc 2023-11-29
sorry forgot upload benchmark result. here it is

```
╰─❯ benchstat /tmp/old-ts.txt /tmp/new2-ts.txt                                                                                                                                                                                                               ─╯
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/trace
cpu: Intel(R) Xeon(R) Platinum 8255C CPU @ 2.50GHz
                                                                   │ /tmp/old-ts.txt │          /tmp/new2-ts.txt           │
                                                                   │     sec/op      │   sec/op     vs base                │
TraceStateParse/duplicate_with_the_same_value-16                        2072.0n ± 1%   478.3n ± 1%  -76.92% (p=0.000 n=10)
TraceStateParse/duplicate_with_different_values-16                      2084.0n ± 1%   480.8n ± 1%  -76.93% (p=0.000 n=10)
TraceStateParse/improperly_formatted_key/value_pair-16                  1109.0n ± 1%   642.4n ± 1%  -42.07% (p=0.000 n=10)
TraceStateParse/upper_case_key-16                                        832.4n ± 2%   632.4n ± 1%  -24.02% (p=0.000 n=10)
TraceStateParse/key_with_invalid_character-16                           1100.0n ± 0%   637.5n ± 1%  -42.05% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_empty_tenant_key-16             1149.0n ± 1%   636.8n ± 0%  -44.57% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_only_tenant-16                   861.9n ± 1%   627.8n ± 0%  -27.17% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_double_tenant_separator-16      1148.0n ± 1%   637.5n ± 0%  -44.47% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_multiple_tenants-16             1268.0n ± 1%   644.4n ± 0%  -49.18% (p=0.000 n=10)
TraceStateParse/key_too_long-16                                         7196.0n ± 1%   911.1n ± 0%  -87.34% (p=0.000 n=10)
TraceStateParse/key_too_long,_with_tenant-16                           28267.0n ± 1%   904.9n ± 0%  -96.80% (p=0.000 n=10)
TraceStateParse/tenant_too_long-16                                      3297.5n ± 3%   789.6n ± 0%  -76.05% (p=0.000 n=10)
TraceStateParse/multiple_values_for_a_single_key-16                     1352.0n ± 1%   640.2n ± 1%  -52.64% (p=0.000 n=10)
TraceStateParse/no_value-16                                             1144.5n ± 1%   623.4n ± 1%  -45.53% (p=0.000 n=10)
TraceStateParse/too_many_members-16                                     37.155µ ± 2%   7.835µ ± 2%  -78.91% (p=0.000 n=10)
TraceStateParse/valid_key/value_list-16                                 5547.5n ± 1%   272.4n ± 0%  -95.09% (p=0.000 n=10)
TraceStateParse/valid_key/value_list_with_tenant-16                     8599.0n ± 1%   292.3n ± 0%  -96.60% (p=0.000 n=10)
TraceStateParse/empty_input-16                                           2.289n ± 1%   2.285n ± 0%        ~ (p=0.402 n=10)
TraceStateParse/single_key_and_value-16                                  944.5n ± 0%   142.1n ± 1%  -84.95% (p=0.000 n=10)
TraceStateParse/single_key_and_value_with_empty_separator-16             973.8n ± 0%   141.1n ± 1%  -85.51% (p=0.000 n=10)
TraceStateParse/multiple_keys_and_values-16                             1888.5n ± 1%   303.2n ± 0%  -83.94% (p=0.000 n=10)
TraceStateParse/with_a_key_at_maximum_length-16                        12882.5n ± 1%   593.9n ± 0%  -95.39% (p=0.000 n=10)
TraceStateParse/with_a_key_and_tenant_at_maximum_length-16             27831.5n ± 1%   605.2n ± 0%  -97.83% (p=0.000 n=10)
TraceStateParse/with_maximum_members-16                                 34.634µ ± 1%   6.802µ ± 1%  -80.36% (p=0.000 n=10)
TraceStateParse/with_several_members-16                                 4938.5n ± 1%   773.0n ± 0%  -84.35% (p=0.000 n=10)
TraceStateParse/with_tabs_between_members-16                            2954.5n ± 0%   496.8n ± 1%  -83.19% (p=0.000 n=10)
TraceStateParse/with_multiple_tabs_between_members-16                   2950.5n ± 2%   494.6n ± 0%  -83.24% (p=0.000 n=10)
TraceStateParse/with_space_at_the_end_of_the_member-16                   963.4n ± 1%   139.9n ± 0%  -85.48% (p=0.000 n=10)
TraceStateParse/with_tab_at_the_end_of_the_member-16                     935.5n ± 1%   139.7n ± 1%  -85.07% (p=0.000 n=10)
TraceStateParse/with_tab_and_space_at_the_end_of_the_member-16           974.9n ± 1%   140.7n ± 1%  -85.57% (p=0.000 n=10)
TraceStateString/valid_key/value_list-16                                 280.1n ± 1%   104.0n ± 0%  -62.87% (p=0.000 n=10)
TraceStateString/valid_key/value_list_with_tenant-16                     283.3n ± 1%   109.9n ± 1%  -61.21% (p=0.000 n=10)
TraceStateString/single_key_and_value-16                                246.45n ± 1%   33.55n ± 0%  -86.39% (p=0.000 n=10)
TraceStateString/single_key_and_value_with_empty_separator-16           244.60n ± 1%   33.55n ± 0%  -86.29% (p=0.000 n=10)
TraceStateString/multiple_keys_and_values-16                            536.55n ± 1%   78.38n ± 0%  -85.39% (p=0.000 n=10)
TraceStateString/with_a_key_at_maximum_length-16                         667.9n ± 1%   131.9n ± 1%  -80.25% (p=0.000 n=10)
TraceStateString/with_a_key_and_tenant_at_maximum_length-16              666.9n ± 1%   131.5n ± 0%  -80.28% (p=0.000 n=10)
TraceStateString/with_maximum_members-16                                7215.0n ± 1%   781.8n ± 1%  -89.16% (p=0.000 n=10)
TraceStateString/with_several_members-16                                1196.0n ± 1%   157.2n ± 0%  -86.86% (p=0.000 n=10)
TraceStateString/with_tabs_between_members-16                            745.6n ± 1%   132.9n ± 1%  -82.17% (p=0.000 n=10)
TraceStateString/with_multiple_tabs_between_members-16                   747.9n ± 1%   132.5n ± 0%  -82.28% (p=0.000 n=10)
TraceStateString/with_space_at_the_end_of_the_member-16                 245.25n ± 1%   33.56n ± 0%  -86.31% (p=0.000 n=10)
TraceStateString/with_tab_at_the_end_of_the_member-16                   244.75n ± 2%   33.51n ± 1%  -86.31% (p=0.000 n=10)
TraceStateString/with_tab_and_space_at_the_end_of_the_member-16         242.75n ± 1%   33.49n ± 0%  -86.20% (p=0.000 n=10)
TraceStateInsert/add_new-16                                              916.2n ± 1%   122.4n ± 0%  -86.64% (p=0.000 n=10)
TraceStateInsert/replace-16                                              491.4n ± 1%   126.8n ± 0%  -74.20% (p=0.000 n=10)
TraceStateInsert/invalid_key-16                                         584.40n ± 1%   20.14n ± 0%  -96.55% (p=0.000 n=10)
TraceStateInsert/invalid_value-16                                       627.70n ± 0%   19.86n ± 1%  -96.84% (p=0.000 n=10)
TraceStateInsert/invalid_key/value-16                                   587.20n ± 2%   20.16n ± 1%  -96.57% (p=0.000 n=10)
TraceStateInsert/drop_the_right-most_member(oldest)_in_queue-16          778.4n ± 1%   387.8n ± 1%  -50.18% (p=0.000 n=10)
geomean                                                                  1.266µ        221.7n       -82.49%

                                                                   │ /tmp/old-ts.txt │             /tmp/new2-ts.txt              │
                                                                   │      B/op       │     B/op      vs base                     │
TraceStateParse/duplicate_with_the_same_value-16                        387.0 ± 0%       128.0 ± 0%   -66.93% (p=0.000 n=10)
TraceStateParse/duplicate_with_different_values-16                      387.0 ± 0%       128.0 ± 0%   -66.93% (p=0.000 n=10)
TraceStateParse/improperly_formatted_key/value_pair-16                  226.0 ± 0%       208.0 ± 0%    -7.96% (p=0.000 n=10)
TraceStateParse/upper_case_key-16                                       226.0 ± 0%       208.0 ± 0%    -7.96% (p=0.000 n=10)
TraceStateParse/key_with_invalid_character-16                           226.0 ± 0%       208.0 ± 0%    -7.96% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_empty_tenant_key-16             242.0 ± 0%       208.0 ± 0%   -14.05% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_only_tenant-16                  243.0 ± 0%       208.0 ± 0%   -14.40% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_double_tenant_separator-16      242.0 ± 0%       208.0 ± 0%   -14.05% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_multiple_tenants-16             242.0 ± 0%       208.0 ± 0%   -14.05% (p=0.000 n=10)
TraceStateParse/key_too_long-16                                        1358.0 ± 0%       752.0 ± 0%   -44.62% (p=0.000 n=10)
TraceStateParse/key_too_long,_with_tenant-16                           1325.0 ± 0%       720.0 ± 0%   -45.66% (p=0.000 n=10)
TraceStateParse/tenant_too_long-16                                      614.0 ± 0%       256.0 ± 0%   -58.31% (p=0.000 n=10)
TraceStateParse/multiple_values_for_a_single_key-16                     226.0 ± 0%       208.0 ± 0%    -7.96% (p=0.000 n=10)
TraceStateParse/no_value-16                                             226.0 ± 0%       192.0 ± 0%   -15.04% (p=0.000 n=10)
TraceStateParse/too_many_members-16                                  10.471Ki ± 0%     6.193Ki ± 0%   -40.85% (p=0.000 n=10)
TraceStateParse/valid_key/value_list-16                                161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/valid_key/value_list_with_tenant-16                    161.00 ± 1%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/empty_input-16                                          0.000 ± 0%       0.000 ± 0%         ~ (p=1.000 n=10) ¹
TraceStateParse/single_key_and_value-16                                161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/single_key_and_value_with_empty_separator-16           177.00 ± 0%       32.00 ± 0%   -81.92% (p=0.000 n=10)
TraceStateParse/multiple_keys_and_values-16                            355.00 ± 0%       96.00 ± 0%   -72.96% (p=0.000 n=10)
TraceStateParse/with_a_key_at_maximum_length-16                        355.00 ± 0%       96.00 ± 0%   -72.96% (p=0.000 n=10)
TraceStateParse/with_a_key_and_tenant_at_maximum_length-16             355.00 ± 0%       96.00 ± 0%   -72.96% (p=0.000 n=10)
TraceStateParse/with_maximum_members-16                               8.170Ki ± 0%     4.092Ki ± 0%   -49.92% (p=0.000 n=10)
TraceStateParse/with_several_members-16                                1129.5 ± 0%       480.0 ± 0%   -57.50% (p=0.000 n=10)
TraceStateParse/with_tabs_between_members-16                            613.0 ± 0%       224.0 ± 0%   -63.46% (p=0.000 n=10)
TraceStateParse/with_multiple_tabs_between_members-16                   613.0 ± 0%       224.0 ± 0%   -63.46% (p=0.000 n=10)
TraceStateParse/with_space_at_the_end_of_the_member-16                 161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/with_tab_at_the_end_of_the_member-16                   161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateParse/with_tab_and_space_at_the_end_of_the_member-16         161.00 ± 0%       32.00 ± 0%   -80.12% (p=0.000 n=10)
TraceStateString/valid_key/value_list-16                                192.0 ± 0%       192.0 ± 0%         ~ (p=1.000 n=10) ¹
TraceStateString/valid_key/value_list_with_tenant-16                    208.0 ± 0%       224.0 ± 0%    +7.69% (p=0.000 n=10)
TraceStateString/single_key_and_value-16                               53.000 ± 0%       8.000 ± 0%   -84.91% (p=0.000 n=10)
TraceStateString/single_key_and_value_with_empty_separator-16          53.000 ± 0%       8.000 ± 0%   -84.91% (p=0.000 n=10)
TraceStateString/multiple_keys_and_values-16                           120.00 ± 0%       24.00 ± 0%   -80.00% (p=0.000 n=10)
TraceStateString/with_a_key_at_maximum_length-16                        677.0 ± 0%       296.0 ± 0%   -56.28% (p=0.000 n=10)
TraceStateString/with_a_key_and_tenant_at_maximum_length-16             677.0 ± 0%       296.0 ± 0%   -56.28% (p=0.000 n=10)
TraceStateString/with_maximum_members-16                               2081.0 ± 0%      1016.0 ± 0%   -51.18% (p=0.000 n=10)
TraceStateString/with_several_members-16                               304.00 ± 0%       56.00 ± 0%   -81.58% (p=0.000 n=10)
TraceStateString/with_tabs_between_members-16                          184.00 ± 0%       56.00 ± 0%   -69.57% (p=0.000 n=10)
TraceStateString/with_multiple_tabs_between_members-16                 184.00 ± 0%       56.00 ± 0%   -69.57% (p=0.000 n=10)
TraceStateString/with_space_at_the_end_of_the_member-16                53.000 ± 0%       8.000 ± 0%   -84.91% (p=0.000 n=10)
TraceStateString/with_tab_at_the_end_of_the_member-16                  53.000 ± 0%       8.000 ± 0%   -84.91% (p=0.000 n=10)
TraceStateString/with_tab_and_space_at_the_end_of_the_member-16        53.000 ± 0%       8.000 ± 0%   -84.91% (p=0.000 n=10)
TraceStateInsert/add_new-16                                             291.0 ± 0%       128.0 ± 0%   -56.01% (p=0.000 n=10)
TraceStateInsert/replace-16                                             96.00 ± 1%       96.00 ± 0%         ~ (p=0.211 n=10)
TraceStateInsert/invalid_key-16                                         80.00 ± 0%        0.00 ± 0%  -100.00% (p=0.000 n=10)
TraceStateInsert/invalid_value-16                                       80.00 ± 0%        0.00 ± 0%  -100.00% (p=0.000 n=10)
TraceStateInsert/invalid_key/value-16                                   80.00 ± 0%        0.00 ± 0%  -100.00% (p=0.000 n=10)
TraceStateInsert/drop_the_right-most_member(oldest)_in_queue-16       1.012Ki ± 0%     1.000Ki ± 0%    -1.16% (p=0.000 n=10)
geomean                                                                            ²                 ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                                                                   │ /tmp/old-ts.txt │            /tmp/new2-ts.txt             │
                                                                   │    allocs/op    │ allocs/op   vs base                     │
TraceStateParse/duplicate_with_the_same_value-16                        8.000 ± 0%     3.000 ± 0%   -62.50% (p=0.000 n=10)
TraceStateParse/duplicate_with_different_values-16                      8.000 ± 0%     3.000 ± 0%   -62.50% (p=0.000 n=10)
TraceStateParse/improperly_formatted_key/value_pair-16                  6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/upper_case_key-16                                       6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/key_with_invalid_character-16                           6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_empty_tenant_key-16             6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_only_tenant-16                  6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_double_tenant_separator-16      6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/multiple_keys,_one_with_multiple_tenants-16             6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/key_too_long-16                                        14.000 ± 0%     6.000 ± 0%   -57.14% (p=0.000 n=10)
TraceStateParse/key_too_long,_with_tenant-16                           14.000 ± 0%     6.000 ± 0%   -57.14% (p=0.000 n=10)
TraceStateParse/tenant_too_long-16                                     14.000 ± 0%     6.000 ± 0%   -57.14% (p=0.000 n=10)
TraceStateParse/multiple_values_for_a_single_key-16                     6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/no_value-16                                             6.000 ± 0%     5.000 ± 0%   -16.67% (p=0.000 n=10)
TraceStateParse/too_many_members-16                                     80.00 ± 0%     13.00 ± 0%   -83.75% (p=0.000 n=10)
TraceStateParse/valid_key/value_list-16                                 4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/valid_key/value_list_with_tenant-16                     4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/empty_input-16                                          0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
TraceStateParse/single_key_and_value-16                                 4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/single_key_and_value_with_empty_separator-16            4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/multiple_keys_and_values-16                             7.000 ± 0%     2.000 ± 0%   -71.43% (p=0.000 n=10)
TraceStateParse/with_a_key_at_maximum_length-16                         7.000 ± 0%     2.000 ± 0%   -71.43% (p=0.000 n=10)
TraceStateParse/with_a_key_and_tenant_at_maximum_length-16              7.000 ± 0%     2.000 ± 0%   -71.43% (p=0.000 n=10)
TraceStateParse/with_maximum_members-16                                 75.00 ± 0%     10.00 ± 0%   -86.67% (p=0.000 n=10)
TraceStateParse/with_several_members-16                                15.000 ± 0%     4.000 ± 0%   -73.33% (p=0.000 n=10)
TraceStateParse/with_tabs_between_members-16                           10.000 ± 0%     3.000 ± 0%   -70.00% (p=0.000 n=10)
TraceStateParse/with_multiple_tabs_between_members-16                  10.000 ± 0%     3.000 ± 0%   -70.00% (p=0.000 n=10)
TraceStateParse/with_space_at_the_end_of_the_member-16                  4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/with_tab_at_the_end_of_the_member-16                    4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateParse/with_tab_and_space_at_the_end_of_the_member-16          4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/valid_key/value_list-16                                4.000 ± 0%     2.000 ± 0%   -50.00% (p=0.000 n=10)
TraceStateString/valid_key/value_list_with_tenant-16                    4.000 ± 0%     2.000 ± 0%   -50.00% (p=0.000 n=10)
TraceStateString/single_key_and_value-16                                4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/single_key_and_value_with_empty_separator-16           4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/multiple_keys_and_values-16                            8.000 ± 0%     2.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/with_a_key_at_maximum_length-16                        8.000 ± 0%     2.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/with_a_key_and_tenant_at_maximum_length-16             8.000 ± 0%     2.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/with_maximum_members-16                               98.000 ± 0%     7.000 ± 0%   -92.86% (p=0.000 n=10)
TraceStateString/with_several_members-16                               17.000 ± 0%     3.000 ± 0%   -82.35% (p=0.000 n=10)
TraceStateString/with_tabs_between_members-16                          11.000 ± 0%     3.000 ± 0%   -72.73% (p=0.000 n=10)
TraceStateString/with_multiple_tabs_between_members-16                 11.000 ± 0%     3.000 ± 0%   -72.73% (p=0.000 n=10)
TraceStateString/with_space_at_the_end_of_the_member-16                 4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/with_tab_at_the_end_of_the_member-16                   4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateString/with_tab_and_space_at_the_end_of_the_member-16         4.000 ± 0%     1.000 ± 0%   -75.00% (p=0.000 n=10)
TraceStateInsert/add_new-16                                             2.000 ± 0%     1.000 ± 0%   -50.00% (p=0.000 n=10)
TraceStateInsert/replace-16                                             1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=10) ¹
TraceStateInsert/invalid_key-16                                         3.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
…